refactor rooms, and move services over
