+++ src/contexts/WargameProvider.tsx
@@ -1,14 +1,12 @@
 import { useState, ReactNode, useMemo } from 'react'
 import { WargameContext } from './WargameContext'
-import { RoomType } from '../types/rooms'
-import { XMPPService } from '../rooms-api/xmpp/XMPPService'
+import { XMPPService } from '../services/XMPPService'
 
 interface WargameProviderProps {
   children: ReactNode
 }
 
 export const WargameProvider = ({ children }: WargameProviderProps) => {
-  const [rooms, setRooms] = useState<RoomType[]>([])
   const [xmppClient, setXmppClient] = useState<XMPPService | null | undefined>(undefined) 
   
   const loggedIn = useMemo(() => {
@@ -17,8 +15,6 @@ export const WargameProvider = ({ children }: WargameProviderProps) => {
 
   const value = {
     loggedIn,
-    rooms,
-    setRooms,
     xmppClient,
     setXmppClient
   }
