+++ src/rooms-test/xmpp/xmppMUC.test.ts
@@ -1,6 +1,6 @@
-import { XMPPService } from '../../rooms-api/xmpp/XMPPService.js'
+import { XMPPService } from '../../services/XMPPService.js'
 import { loadOpenfireConfig } from '../../utils/config.js'
-import { RoomMessage } from '../../rooms-api/xmpp/types.js'
+import { RoomMessage } from '../../services/types.js'
 
 describe('XMPP MUC (Multi-User Chat)', () => {
   let xmppService: XMPPService
