+++ src/services/XMPPService.ts
